Improve EntityTracker for multiple spawn/despawn as well as Self.
Track Self entities in the EntityTracker with all the other entities.
Move handling of kEntityMovementTimerEnded event into the MobileEntity.
Call Bot::onUpdate for all events; remove a lot of special handler functions which no longer do anything.
Create/destroy AutoPotion state machine on self spawn/despawn.
Fix bug where MobileEntity::moving_ was uninitialized.
Add entity ID to entered new region event.
Add a few missing strings for event names.
Print thread ID when logging packets.
Handle multiple characters seeing/"tracking" the same entity.
Add checks in Session for initialization.

+++ Hyperbot/src/ui/userInterface.cpp
@@ -327,7 +327,7 @@ void UserInterface::handleRequest(const zmq::message_t &request) {
     case proto::request::RequestMessage::BodyCase::kPacketData: {
         const proto::request::PacketToInject &packet = requestMsg.packetdata();
         const event::InjectPacket::Direction dir = (packet.direction() == proto::request::PacketToInject::kClientToServer) ? event::InjectPacket::Direction::kClientToServer : event::InjectPacket::Direction::kServerToClient;
-        eventBroker_.publishEvent<event::InjectPacket>(dir, packet.opcode(), packet.data());
+        eventBroker_.publishEvent<event::InjectPacket>(dir, static_cast<uint16_t>(packet.opcode()), packet.data());
         break;
       }
     case proto::request::RequestMessage::BodyCase::kDoAction: {
